requirements.txt
PyYAML
PyContracts
ConfTools
RepRep
compmake
comptests
SystemCmd
hdflog
